requirements.txt
flask
torch 
torchvision 
pillow
ipykernel
matplotlib
numpy
pandas
asgiref